Error fix 2

+++ qBitrr/arss.py
@@ -3536,10 +3536,11 @@ class Arr:
                 entry["episodeId"] for entry in self.queue if entry.get("episodeId")
             }
             self.logger.debug("Test 3")
-            self.model_queue.delete().where(
-                self.model_queue.EntryId.not_in(list(self.queue_file_ids))
-            )
-            self.logger.debug("Test 7")
+            if self.model_queue:
+                self.model_queue.delete().where(
+                    self.model_queue.EntryId.not_in(list(self.queue_file_ids))
+                )
+                self.logger.debug("Test 7")
         elif self.type == "radarr":
             self.logger.debug("Test 5")
             self.requeue_cache = {
@@ -3550,10 +3551,11 @@ class Arr:
                 entry["movieId"] for entry in self.queue if entry.get("movieId")
             }
             self.logger.debug("Test 3")
-            self.model_queue.delete().where(
-                self.model_queue.EntryId.not_in(list(self.queue_file_ids))
-            )
-            self.logger.debug("Test 7")
+            if self.model_queue:
+                self.model_queue.delete().where(
+                    self.model_queue.EntryId.not_in(list(self.queue_file_ids))
+                )
+                self.logger.debug("Test 7")
 
         self._update_bad_queue_items()
